Major Additions
Created a system to:
-procedurally generate data for enemies on an active "block" by pulling from seed data for the block and combining with default enemy data from an enemy data defaults spreadsheet
-load and unload enemy data from active memory and hard drive

this results in procedural enemy spawning, the capacity for persistent enemies (they dont dissapear once you leave an area, as their data is saved to the hard drive), and ultimately lays the groundwork other open world components. eg. this setup can be replicated for foliage, items, and anything else that should be persistent, dynamic, and procedural